# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Healthcare Staffing Platform API Doc
  description: |
    On-Demand Healthcare Staffing & Visit Management Portal — REST API.

    ## Authentication
    All protected endpoints require a **Bearer JWT** in the `Authorization` header:
    ```
    Authorization: Bearer <accessToken>
    ```
    Obtain tokens via `POST /auth/login`. Access tokens expire in **15 minutes**;
    use `POST /auth/refresh` with a valid refresh token to rotate.

    ## Roles
    | Role | Description |
    |------|-------------|
    | `SUPER_ADMIN` | Full platform control |
    | `RECRUITER` | Read-only admin access |
    | `FACILITY_ADMIN` | Manages one facility tenant |
    | `TEAM_MEMBER` | Scoped facility operator |
    | `NURSE` | Field professional (mobile) |

    ## File Uploads
    All uploads are stored in **DigitalOcean Spaces** (private).
    Use `POST /storage/signed-upload-url` for direct client-to-Spaces uploads,
    or `POST /storage/upload/{folder}` for server-proxied multipart uploads.
    Download links are short-lived signed URLs — never public.

    ## Pagination
    List endpoints accept `page` and `limit` query params and return:
    ```json
    { "pagination": { "page": 1, "limit": 20, "total": 100, "totalPages": 5 } }
    ```

    ## EVV Geofencing
    Check-in/out enforces a **200-metre geofence** around the case address (Haversine).
    Breaches create a `FLAGGED` visit requiring admin override approval.
  version: 1.0.0
  contact:
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── Reusable Parameters ──────────────────
  parameters:
    PageParam:
      name: page
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
    FromParam:
      name: from
      in: query
      description: ISO 8601 start date for range filter
      schema: { type: string, format: date }
    ToParam:
      name: to
      in: query
      description: ISO 8601 end date for range filter
      schema: { type: string, format: date }

  # ─── Reusable Responses ───────────────────
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    InvoiceStatus:
      type: string
      enum: [DRAFT, ISSUED, PAID, OVERDUE, VOID]
    PayoutStatus:
      type: string
      enum: [PENDING, SETTLED, FAILED]
...
    SuccessResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        message: { type: string }
        data:    { type: object }
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, example: 'Resource not found' }
...
    PaginationMeta:
      type: object
      properties:
        page:       { type: integer }
        limit:      { type: integer }
        total:      { type: integer }
        totalPages: { type: integer }
        hasNext:    { type: boolean }
        hasPrev:    { type: boolean }

    # ── Auth ───────────────────────────────
...
    Invoice:
      type: object
      properties:
        id:             { type: string, format: uuid }
        invoiceNumber:  { type: string, example: 'INV-2026-00142' }
        facilityId:     { type: string, format: uuid }
        status:         { $ref: '#/components/schemas/InvoiceStatus' }
        subtotal:       { type: number, format: double }
        tax:            { type: number, format: double }
        total:          { type: number, format: double }
        periodStart:    { type: string, format: date }
        periodEnd:      { type: string, format: date }
        dueAt:          { type: string, format: date-time, nullable: true }
        paidAt:         { type: string, format: date-time, nullable: true }
        fileUrl:        { type: string, nullable: true }
        lineItems:
          type: array
          items:
            type: object
            properties:
              description: { type: string }
              quantity:    { type: number }
              unitRate:    { type: number }
              amount:      { type: number }
...
    Payout:
      type: object
      properties:
        id:               { type: string, format: uuid }
        nurseProfileId:   { type: string, format: uuid }
        grossCharge:      { type: number, format: double }
        netPayout:        { type: number, format: double }
        systemCommission: { type: number, format: double }
        status:           { $ref: '#/components/schemas/PayoutStatus' }
        paidAt:           { type: string, format: date-time, nullable: true }
        createdAt:        { type: string, format: date-time }

    # ── Storage ────────────────────────────
...
security:
  - BearerAuth: []

# docs/paths-part1.yaml  (auth · users · credentials · facilities)
paths:
...
  /billing/invoices:
    post:
      tags: [Billing]
      summary: Generate an invoice for a facility
      description: '**Roles:** SUPER_ADMIN, FACILITY_ADMIN'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [facilityId, periodStart, periodEnd]
              properties:
                facilityId:  { type: string, format: uuid }
                periodStart: { type: string, format: date }
                periodEnd:   { type: string, format: date }
                notes:       { type: string }
                dueAt:       { type: string, format: date-time }
      responses:
        '201':
          description: Invoice generated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Invoice' }

    get:
      tags: [Billing]
      summary: List invoices
      description: Facility users see only their own invoices.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: facilityId
          in: query
          schema: { type: string, format: uuid }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/InvoiceStatus' }
        - $ref: '#/components/parameters/FromParam'
        - $ref: '#/components/parameters/ToParam'
      responses:
        '200':
          description: Paginated invoice list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:    { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Invoice' }
                  pagination: { $ref: '#/components/schemas/PaginationMeta' }
  /billing/invoices/{id}:
    get:
      tags: [Billing]
      summary: Get invoice detail with line items
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Invoice detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Invoice' }
        '404': { $ref: '#/components/responses/NotFound' }
  /billing/invoices/{id}/issue:
    patch:
      tags: [Billing]
      summary: Issue a draft invoice (send to facility)
      description: '**Roles:** SUPER_ADMIN'
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Invoice issued
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Invoice' }
  /billing/invoices/{id}/void:
    patch:
      tags: [Billing]
      summary: Void an invoice
      description: '**Roles:** SUPER_ADMIN'
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Invoice voided
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
...
  /billing/payouts:
    get:
      tags: [Billing]
      summary: List payouts
      description: |
        - **NURSE**: sees own payouts only
        - **SUPER_ADMIN**: sees all payouts, filterable by `nurseProfileId`
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: nurseProfileId
          in: query
          schema: { type: string, format: uuid }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/PayoutStatus' }
        - $ref: '#/components/parameters/FromParam'
        - $ref: '#/components/parameters/ToParam'
      responses:
        '200':
          description: Payout list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:    { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Payout' }
                  pagination: { $ref: '#/components/schemas/PaginationMeta' }
...
  /billing/summary:
    get:
      tags: [Billing]
      summary: Billing summary for a facility or platform
      description: '**Roles:** SUPER_ADMIN, FACILITY_ADMIN'
      parameters:
        - name: facilityId
          in: query
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/FromParam'
        - $ref: '#/components/parameters/ToParam'
      responses:
        '200':
          description: Billing summary
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          upcomingCharges: { type: number }
                          overdueBalance:  { type: number }
                          totalPaid:       { type: number }
                          openInvoices:    { type: integer }

  # ═══════════════════════════════════════════
  # REPORTS
  # ═══════════════════════════════════════════
...